flask_app/requirements.txt
flask
gunicorn
matplotlib
Pillow